renamed components PascalCase
